Create CLI wrapper to allow for mock execution.
